Reorganize primitives 2
